pyproject.toml
[project]
name = "ynab-status"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "ynab>=1.8.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
